fix(chrome): search sits right of the inspector toggle, Finder-styled; durationMs honestly fractional
Search (Daniel, 2026-08-23): the field mounts in the inspector-section
toolbar AFTER the toggle — the only way to sit right of it, since the
content section always renders left — and the grey pill becomes a
Finder-like hairline stroke.

RunComparisonSide.durationMs widens to Double: the engine reports fractional
milliseconds (duration_ms: number in the regen'd spec — the int coercion
500'd every Compare Runs), and the merge gate caught the one client site
still typed Int.

+++ fichero/fichero/Services/RunComparison.swift
@@ -17,7 +17,10 @@ struct RunComparisonSide: Equatable {
     let workflowName: String
     let status: String
     let error: String?
-    let durationMs: Int?
+    // Double, not Int: the engine reports fractional milliseconds
+    // (duration_ms: number since 2026-08-23 — the int coercion 500'd every
+    // Compare Runs before that).
+    let durationMs: Double?
     let artifactCount: Int
     let stepsTotal: Int
     let stepsFailed: Int
